creation of tables and new mvc layers
Create the rest of the database tables, a trigger on the tv show reminder table, the User Tv Show controller, service and repository. I also make a basic integration with the API themoviedb, and then, create the getTvShowReminderResponseDTO endpoint, that will use the getTvShowDetails endpoint to make request to the API and get the full details of the tv show.