Refactored Service Method getTvShowDetails
I created two new methods, where i put the logic to build the complete image URL, and another method where i filter the video results, we are removing everything from the array that its not a trailer, and keeping the rest. Also I built the complete url for each video.