Refactor TvShowDetails Service
Refactored getTvShowDetailsFromAPI, this endpoint will manage to get the details from the api and nothing more. We then created a new endpoint to get the actual details we are going to return to the controller. This endpoint will map the details we get from the api to a new DTO. We rearranged the information and also modified it, by creating complete url images and urltrailesrs.