http, filter and authenticate endpoint
Configured the HTTP security. Then I, Added a filter to intercept every incoming request once and check the jwt in the header, so i can validate them. Also finished the authenticate endpoint. Tested the endpoint, is returning valid token. After that, I tried to access a secured endpoint with this token, and its working.