Delete endpoints
Added a deleteTvShowReminder endpoint and the service logic for this operation. Also i added a trigger in the database that when we delete a reminder, if this reminder has a user tv show associated it, we deleted it. Also i added a deleteUserTvShow endpoint, with delete on cascade in the DB,  we also delete the reminder associated with it